third_party/WebKit/LayoutTests/fast/events/mouse-event-from-touch-source-device-event-sender-expected.txt

Test that synthetic mouse events from touch events should have a sourceDevice and its firesTouchEvents should be true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Sending gestureTap which synthesizes mousemove, mousedown, mouseup and click
mousemove
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
mousedown
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
mouseup
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
click
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
